refactor: create worker folder
