+++ src/worker/complete.ts
@@ -1,0 +1,8 @@
+import type { CompletionList } from 'vscode-languageserver-protocol'
+
+export function doComplete(state: any, document: any): CompletionList | undefined {
+  // eslint-disable-next-line no-console
+  console.log('start complete', document)
+
+  return undefined
+}
